# api/openapi.yaml
openapi: 3.1.0
info:
  title: MemXplore API
  version: 0.1.0
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
  description: >-
    Provider-neutral agent-memory protocol v1. Recall returns structured evidence and
    provenance, never a generated answer. Formation is durable and may be asynchronous.
servers:
  - url: http://127.0.0.1:7878
    description: Default loopback daemon
security:
  - bearerAuth: []
paths:
...
  /v1/remember:
    post:
      operationId: remember
      summary: Capture evidence and enqueue durable formation
      description: Requires memory:write. Idempotency is scoped to the authenticated namespace.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RememberRequest" }
      responses:
        "200":
          description: Formation reached terminal state within the requested wait
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RememberResponse" }
        "202":
          description: Durable job accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RememberResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    BadRequest:
      description: Invalid schema or domain request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Authenticated principal lacks scope or owner access
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    Content:
      type: object
      additionalProperties: false
      required: [parts]
      properties:
        parts:
          type: array
          minItems: 1
          items:
            oneOf:
              - type: object
                additionalProperties: false
                required: [kind, text]
                properties:
                  kind: { const: text }
                  text: { type: string, minLength: 1 }
              - type: object
                additionalProperties: false
                required: [kind, artifact]
                properties:
                  kind: { const: artifact }
                  artifact:
                    type: object
                    additionalProperties: false
                    required: [digest, media_type, size]
                    properties:
                      digest: { type: string, pattern: "^sha256:[0-9a-f]{64}$" }
                      media_type: { type: string }
                      size: { type: integer, minimum: 0 }
    RememberRequest:
      type: object
      additionalProperties: false
      required: [idempotency_key, owner, subject, content, function]
      properties:
        observation_id: { type: string }
        idempotency_key: { type: string, minLength: 1 }
        owner: { type: string, minLength: 1 }
        subject: { type: string, minLength: 1 }
        context: { type: string }
        visibility: { enum: [private, shared, public] }
        source_kind: { type: string, default: api }
        source_reference: { type: string }
        content: { $ref: "#/components/schemas/Content" }
        evidence_class: { enum: [untrusted, asserted, verified, policy] }
        policy_authority: { type: string }
        captured_at: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: { type: string } }
        function: { enum: [factual, experiential, working] }
        strategy: { enum: [generator-free, assisted], default: generator-free }
        working_ttl_seconds: { type: integer, minimum: 1, maximum: 2592000 }
        working_global_recall: { type: boolean, default: false }
        wait_milliseconds: { type: integer, minimum: 0, maximum: 30000 }
    RememberResponse:
      type: object
      required: [job]
      properties:
        job: { $ref: "#/components/schemas/Job" }
    Job:
      type: object
      required: [id, namespace, kind, state, idempotency_key, attempts, available_at, created_at, updated_at]
      properties:
        id: { type: string }
        namespace: { type: string }
        kind: { type: string }
        state: { enum: [queued, running, succeeded, failed, canceled] }
        idempotency_key: { type: string }
        payload: {}
        result: {}
        error_code: { type: string }
        error_message: { type: string }
        attempts: { type: integer }
        available_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}